Refactor code to fit Code Quality standards
Code does not follow Code Quality standards closely.

Code that is written to different standards can make it harder to read
and follow along when being reviewed by peers, hence we should aim to
improve the readability of the code and follow an agreed-upon coding
standard.

Let's refactor the code to follow the SE-EDU Java coding standard.

The SE-EDU Java coding standard is the agreed standard to follow by the
course, so we should follow this instead of relying on our intuition.

Refer to the SE-EDU Java coding standards at
https://se-education.org/guides/conventions/java/intermediate.html