Add ability to prioritise tasks
Tasks were ordered based on order of insertion, and could not be edited
further.

User may have preferences on the ordering of the tasks, thus they may
want to be able to order the tasks according to their preferences.

Let's add a priority attribute to Task objects, and allow for users to
be able to sort the current task list based on this attribute.

Using the comparator interface allows for tasks objects to have a
natural ordering, allowing for easier sorting.

On top of adding sort and priority feature, also modified .gitignore to
stop tracking test txt files.